requirements.txt
# ASPI - Assistente Inteligente para Dados do Setor Elétrico
# Dependências principais

# Framework Web
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Dados e Análise
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Base de Dados
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Cache e Performance
redis>=5.0.0
asyncio-throttle>=1.0.0

# APIs e Requisições
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# Download e Progress
tqdm>=4.66.0

# Processamento de Linguagem Natural
openai>=1.3.0
anthropic>=0.7.0

# Automação e Workflows
# (n8n via Docker)

# Monitoramento e Logs
structlog>=23.2.0
prometheus-client>=0.19.0

# Utilitários
python-dotenv>=1.0.0
pydantic>=2.5.0
click>=8.1.0
rich>=13.7.0

# Desenvolvimento e Testes
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.7.0

# Segurança
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4